backend/django/core/urls.py
from django.contrib import admin
from django.urls import path, include
from drf_spectacular.views import (
    SpectacularAPIView,
    SpectacularRedocView,
    SpectacularSwaggerView,
)
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)


urlpatterns = [
    path("admin/", admin.site.urls),
    path(
        "schema/",
        SpectacularAPIView.as_view(),
        name="schema",
    ),
    path(
        "api/docs/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
    path(
        "api/redoc/",
        SpectacularRedocView.as_view(url_name="schema"),
        name="redoc",
    ),
    path("api/refresh/", TokenRefreshView.as_view(), name="refresh-token",),
    path(
        "api/login/",
        TokenObtainPairView.as_view(),
        name="login",
    ),
    path("api/auth/", include("auth_app.urls"), name="register"),
    path("api/", include("sacco.urls"), name="sacco"),
]
